Refactored these to remove all classes. Classes aren't needed here, functions will work just fine. Send2HumioHEC.py: I removed all logging and moved that over to the main function so this is a more pure function
ProofpointSIEM2Humio_Main: Removed the if statements since we can combine those using the event_types and a loop. Also, this is using the requests.auth.HTTPBasicAuth to create the SIEM API authentication method. This now combines all of the events into a newline delimited list of json objects since Humio can handles data pushes as such.

I also add a way to persist the last time the query sent data for. hardcoding the sinceSeconds=3600 was inflexible so this script can now be installed and run via cron or a timer and will pick up contiguous data

ProofpointSIEM2Humio_Config.py: removed the hardcode params to the SIEM API. Also mapped the tokens for each event_type to make it easier to call